+++ package.json
@@ -5,7 +5,6 @@
     "type": "git",
     "url": ""
   },
-  "type": "commonjs",
   "description": "REST service starter for RS School",
   "main": "src/server.js",
   "scripts": {
@@ -53,6 +52,8 @@
     "yamljs": "0.3.0"
   },
   "devDependencies": {
+    "@types/node": "^16.11.12",
+    "@types/uuid": "^8.3.3",
     "chai": "4.3.4",
     "debug": "4.3.2",
     "dirty-chai": "2.0.1",
@@ -67,6 +68,7 @@
     "lint-staged": "12.0.3",
     "nodemon": "2.0.15",
     "prettier": "2.4.1",
-    "supertest": "6.1.6"
+    "supertest": "6.1.6",
+    "typescript": "^4.5.2"
   }
 }
