--- src/resources/boards/schemas/index.js
@@ -1,7 +1,0 @@
-const board = require('./board.schema.json');
-const boardPost = require('./board-post.schema.json');
-
-module.exports = {
-  board,
-  boardPost
-}
