+++ src/resources/boards/schemas/index.ts
@@ -1,0 +1,7 @@
+import board from './board.schema.json';
+import boardPost from './board-post.schema.json';
+
+export default {
+  board,
+  boardPost
+}
