--- src/resources/columns/schemas/index.js
@@ -1,7 +1,0 @@
-const columnSchema = require('./column.schema.json')
-const columnPostSchema = require('./column-post.schema.json')
-
-module.exports = {
-  columnSchema,
-  columnPostSchema
-}
