+++ src/resources/columns/schemas/index.ts
@@ -1,0 +1,7 @@
+import columnSchema from './column.schema.json'
+import columnPostSchema from './column-post.schema.json'
+
+export default {
+  columnSchema,
+  columnPostSchema
+}
