--- src/resources/tasks/schemas/index.js
@@ -1,7 +1,0 @@
-const task = require('./task.schema.json');
-const taskPost = require('./task-post.schema.json');
-
-module.exports = {
-  task,
-  taskPost
-}
