+++ src/resources/tasks/schemas/index.ts
@@ -1,0 +1,7 @@
+import task from './task.schema.json';
+import taskPost from './task-post.schema.json';
+
+export default {
+  task,
+  taskPost
+}
