--- src/resources/users/schemas/index.js
@@ -1,9 +1,0 @@
-const user = require('./user.schema.json');
-const userGet = require('./user-get.schema.json');
-const userPost = require('./user-post.schema.json');
-
-module.exports = {
-  user,
-  userGet,
-  userPost
-}
