+++ src/resources/users/schemas/index.ts
@@ -1,0 +1,9 @@
+import user from './user.schema.json';
+import userGet from './user-get.schema.json';
+import userPost from './user-post.schema.json';
+
+export default {
+  user,
+  userGet,
+  userPost
+}
