+++ src/server.ts
@@ -1,5 +1,7 @@
-const { PORT } = require('./common/config');
-const app = require('./app');
+import config from './common/config';
+import app from './app';
+
+const { PORT } = config
 
 const start = async () => {
   try {
