--- src/state/index.js
@@ -1,7 +1,0 @@
-const state = {
-  boards: [],
-  users: [],
-  tasks: []
-}
-
-module.exports = state;
