+++ src/state/index.ts
@@ -1,0 +1,9 @@
+import { RootState } from '../types'
+
+const state: RootState = {
+  boards: [],
+  users: [],
+  tasks: []
+}
+
+export default state;
